apps/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
SQLAlchemy==2.0.29
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.7.1
fastapi-users[sqlalchemy]==13.0.0
python-multipart==0.0.9
celery==5.4.0
redis==5.0.4
boto3==1.34.113
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
